Move test_mode_consensus to mock_consensus in the test_journal unit test directory.
