+++ validator/tests/unit3/test_journal/tests.py
@@ -20,28 +20,23 @@ import unittest
 from sawtooth_validator.journal.block_cache import BlockCache
 from sawtooth_validator.journal.block_wrapper import BlockStatus
 from sawtooth_validator.journal.block_wrapper import BlockWrapper
+
 from sawtooth_validator.journal.chain import BlockValidator
 from sawtooth_validator.journal.chain import ChainController
 from sawtooth_validator.journal.journal import Journal
 from sawtooth_validator.journal.publisher import BlockPublisher
-from sawtooth_validator.journal.consensus.test_mode.test_mode_consensus \
-    import \
-    BlockPublisher as TestModePublisher
-from sawtooth_validator.journal.consensus.test_mode.test_mode_consensus \
-    import \
-    BlockVerifier as TestModeVerifier
-from sawtooth_validator.journal.consensus.test_mode \
-    import test_mode_consensus
 from sawtooth_validator.journal.timed_cache import TimedCache
 from sawtooth_validator.protobuf.batch_pb2 import Batch
-from sawtooth_validator.protobuf.block_pb2 import BlockHeader
+
 from test_journal.block_tree_manager import BlockTreeManager
+
 from test_journal.mock import MockBlockSender
 from test_journal.mock import MockNetwork
 from test_journal.mock import MockTransactionExecutor
 from test_journal.mock import SynchronousExecutor
 from test_journal.utils import wait_until
 
+from test_journal import mock_consensus
 
 LOGGER = logging.getLogger(__name__)
 
@@ -72,7 +67,6 @@ class TestBlockCache(unittest.TestCase):
         self.assertTrue("test" in bc)
         self.assertTrue(bc["test"] == "value")
 
-
 class TestBlockPublisher(unittest.TestCase):
     def setUp(self):
         self.blocks = BlockTreeManager()
@@ -82,7 +76,7 @@ class TestBlockPublisher(unittest.TestCase):
 
         LOGGER.info(self.blocks)
         publisher = BlockPublisher(
-            consensus=TestModePublisher(),
+            consensus=mock_consensus.BlockPublisher(),
             transaction_executor=MockTransactionExecutor(),
             block_sender=self.block_sender,
             squash_handler=None,
@@ -115,7 +109,7 @@ class TestBlockValidator(unittest.TestCase):
 
     def create_block_validator(self, new_block, on_block_validated):
         return BlockValidator(
-            consensus=TestModeVerifier(),
+            consensus=mock_consensus.BlockVerifier(),
             new_block=new_block,
             chain_head=self.btm.chain_head,
             block_cache=self.btm.block_cache,
@@ -327,7 +321,7 @@ class TestChainController(unittest.TestCase):
             pass
 
         self.chain_ctrl = ChainController(
-            consensus=TestModeVerifier(),
+            consensus=mock_consensus.BlockVerifier(),
             block_cache=self.blocks.block_cache,
             block_sender=self.block_sender,
             executor=self.executor,
@@ -424,7 +418,7 @@ class TestJournal(unittest.TestCase):
         journal = None
         try:
             journal = Journal(
-                consensus=test_mode_consensus,
+                consensus=mock_consensus,
                 block_store=btm.block_store.store,
                 block_cache=btm.block_cache,
                 block_sender=self.block_sender,
